Big modification - Move the component folder out of app fodler
