+++ components/icons/AgreementIcon.tsx
