+++ components/icons/MobileAppIcon.tsx
