+++ components/icons/NetworkIcon.tsx
