+++ components/icons/SearchIcon.tsx
